# docs/swagger.yaml
openapi: 3.0.1
info:
  title: OoBDev.WebApi
  version: 1.0.0.0
paths:
...
  /api/Ollama:
    post:
      tags:
        - Ollama
      summary: Generate an LLM Response based on the prompt and user input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/json:
              schema:
                type: string
      x-permissions:
        anonymous: true
    get:
      tags:
        - Ollama
      summary: Retrieves the embedding vector for the given text.
      parameters:
        - name: text
          in: query
          description: The text for which to retrieve the embedding vector.
          schema:
            type: string
        - name: model
          in: query
          description: The model for which to retrieve the embedding vector.
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/System.ReadOnlyMemory-System.Single'
            application/json:
              schema:
                $ref: '#/components/schemas/System.ReadOnlyMemory-System.Single'
            text/json:
              schema:
                $ref: '#/components/schemas/System.ReadOnlyMemory-System.Single'
      x-permissions:
        anonymous: false
  /api/Ollama/Streamed:
    post:
      tags:
        - Ollama
      summary: Generate an AbstractAI Streamed Response based on the prompt and user input
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OoBDev.WebApi.Models.GenerativeAiRequestModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  type: string
            application/json:
              schema:
                type: array
                items:
                  type: string
            text/json:
              schema:
                type: array
                items:
                  type: string
      x-permissions:
        anonymous: true
  /api/Ollama/Completion:
    post:
      tags:
        - Ollama
      summary: executes a completion request
      requestBody:
        description: completion request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OoBDev.AI.Models.CompletionRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/OoBDev.AI.Models.CompletionRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/OoBDev.AI.Models.CompletionRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OoBDev.AI.Models.CompletionResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OoBDev.AI.Models.CompletionResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OoBDev.AI.Models.CompletionResponse'
      x-permissions:
        anonymous: false
...
components:
  schemas:
    OoBDev.AI.Models.CompletionRequest:
      required:
        - prompt
      type: object
      properties:
        model:
          type: string
          description: Gets or initializes the model for the completion request.
          nullable: true
        context:
          type: array
          items:
            type: integer
            format: int64
          description: Gets or initializes the context for the completion request.
          nullable: true
        images:
          type: array
          items:
            type: string
          description: Gets or initializes the images for the completion request.
          nullable: true
        prompt:
          type: string
          description: Gets or initializes the prompt for the completion request.
          nullable: true
        system:
          type: string
          description: Gets or initializes the system for the completion request.
          nullable: true
        template:
          type: string
          description: Gets or initializes the template for the completion request.
          nullable: true
      additionalProperties: false
      description: Represents a completion request.
    OoBDev.AI.Models.CompletionResponse:
      required:
        - context
        - response
      type: object
      properties:
        context:
          type: array
          items:
            type: integer
            format: int64
          description: Gets or sets the context for the completion response.
          nullable: true
        response:
          type: string
          description: Gets or sets the response generated by the completion process.
          nullable: true
      additionalProperties: false
      description: Represents a completion response.
...
    OoBDev.WebApi.Models.GenerativeAiRequestModel:
      required:
        - data
        - promptDetails
        - userInput
      type: object
      properties:
        promptDetails:
          type: string
          description: Gets or sets the prompt details.
          nullable: true
        data:
          type: string
          description: Gets or sets the prompt details.
          nullable: true
        userInput:
          type: string
          description: Gets or sets the user input.
          nullable: true
      additionalProperties: false
      description: Represents the model used for generating AI responses.
    System.ReadOnlyMemory-System.Single:
      type: object
      properties:
        length:
          type: integer
          format: int32
          readOnly: true
        isEmpty:
          type: boolean
          readOnly: true
        span:
          $ref: '#/components/schemas/System.ReadOnlySpan-System.Single'
      additionalProperties: false
    System.ReadOnlySpan-System.Single:
      type: object
      properties:
        length:
          type: integer
          format: int32
          readOnly: true
        isEmpty:
          type: boolean
          readOnly: true
      additionalProperties: false
...